refactor-settings * Move functionality in useFirebaseSettings.ts into SettingsProvider.tsx. * Add new return-statements for SettingsProvider when on a path without a documentId or when settings are being fetched from firebase. * Make sure that SettingsContext.Provider is never given a null value * Change return type of useSettings to never be null
